RF: move split_ext under .support.path
It was introduced in addurls to mimic the way annex splits out extensions.
I think we better adopt it generally, thus moved it into the path module
