+++ datalad/support/tests/test_path.py
@@ -12,6 +12,7 @@ from ..path import (
     abspath,
     curdir,
     robust_abspath,
+    split_ext,
 )
 from ...dochelpers import exc_str
 from ...utils import (
@@ -41,3 +42,17 @@ def test_robust_abspath(tdir):
 
         assert_raises(OSError, abspath, curdir)
         eq_(robust_abspath(curdir), tdir)
+
+
+def test_split_ext():
+    eq_(split_ext("file"), ("file", ""))
+
+    eq_(split_ext("file.py"), ("file", ".py"))
+    eq_(split_ext("file.tar.gz"), ("file", ".tar.gz"))
+    eq_(split_ext("file.toolong.gz"), ("file.toolong", ".gz"))
+
+    eq_(split_ext("file.a.b.c.d"), ("file", ".a.b.c.d"))
+    eq_(split_ext("file.a.b.cccc.d"), ("file", ".a.b.cccc.d"))
+    eq_(split_ext("file.a.b.ccccc.d"), ("file.a.b.ccccc", ".d"))
+
+    eq_(split_ext("file.a.b..c"), ("file", ".a.b..c"))
